Don't set data_valid=1 only because need_csum==0
Setting dava_valid=1 has stronger implications than having need_csum==0:
Former is a _promise_ to the receiving peer that the device actually
_did_ validate the checksum (or calculated it itself, which basically
has the same effect), where the latter does not tell much about the
actual _validity_ of the packet's checksum (it's more a statement about
the sending peer's understanding of the world, which might be false).

Thus, setting data_valid=1 might be a blatant lie if we didn't actually
verify the data in some way.  Furthermore, setting this bit only is
valid for a receiving peer that has checksum offloading negotiated, thus
setting it uncounditionally is a further violation of the spec.

Change-Id: I88a1c1f54b86bdd150de95678d64c41906d2605b