Add method to notify virtio driver of config change
The existing driver notification functionality is not available to the base
L4virtio server class. The virtual method trigger_driver_config_irq() is added
to L4virtio server to allow common code in the base class to trigger a device
configuration change notification. This is necessary for example when the
device sets the DEVICE_NEEDS_RESET status bit.

Change-Id: I27286e1238239b20b940433ec99dd4cb5b460f90